refactor: update docs folder
Closes #49
