--- docs/public/wasm/snippets/dioxus-cli-config-c2775606adf80e62/inline0.js
@@ -1,9 +1,0 @@
-
-        export function getMetaContents(meta_name) {
-            const selector = document.querySelector(`meta[name="${meta_name}"]`);
-            if (!selector) {
-                return null;
-            }
-            return selector.content;
-        }
-
