+++ docs/tailwind_input.css
@@ -1,0 +1,1 @@
+@tailwind base;@tailwind components;@tailwind utilities;@apply py-6;@apply h-1 border-none bg-neutral-300;@apply text-orange-500;@apply text-sky-300;@apply text-orange-300;@apply text-violet-400;@apply text-emerald-700;@apply text-amber-400;@apply text-purple-400;@apply text-blue-400;@apply text-red-400;@apply text-green-400;@apply text-lime-600;@apply text-pink-300;
