+++ docs/wasm/snippets/dioxus-cli-config-c2775606adf80e62/inline0.js
@@ -1,0 +1,9 @@
+
+        export function getMetaContents(meta_name) {
+            const selector = document.querySelector(`meta[name="${meta_name}"]`);
+            if (!selector) {
+                return null;
+            }
+            return selector.content;
+        }
+
