+++ docs/wasm/snippets/dioxus-web-fe956e4b19891335/inline0.js
@@ -1,0 +1,12 @@
+
+export function get_select_data(select) {
+    let values = [];
+    for (let i = 0; i < select.options.length; i++) {
+      let option = select.options[i];
+      if (option.selected) {
+        values.push(option.value.toString());
+      }
+    }
+
+    return values;
+}
